feat: implements AWS SES notifications

+++ app/controllers/product.py
@@ -1,14 +1,16 @@
 import uuid
-from pydantic import validate_email
 from app.repositories.product import ProductRepository
 from app.repositories.user import UserRepository
 from app.structures.product import ProductData
-from werkzeug.security import generate_password_hash
-
 from app.structures.user import Roles
+from app.utils.email_helper import EmailHelper
 
 
 class ProductAlreadyExists(Exception):
+    pass
+
+
+class ProductNotFound(Exception):
     pass
 
 
@@ -37,7 +39,12 @@ class ProductController():
         return product
 
     @classmethod
-    def update_product(cls, public_id: str, name: str, sku: str, price: float, brand: str, queries: int) -> ProductData:
+    def update_product(cls, user_public_id: str, public_id: str, name: str, sku: str, price: float, brand: str) -> ProductData:
+
+        old_product = ProductController.get_product_by_public_id(public_id)
+
+        if not old_product:
+            raise ProductNotFound()
 
         product = ProductData(
             public_id=public_id,
@@ -45,7 +52,7 @@ class ProductController():
             sku=sku,
             price=price,
             brand=brand,
-            queries=queries
+            queries=old_product.queries
         )
 
         if ProductRepository.name_exists(name, public_id):
@@ -55,6 +62,14 @@ class ProductController():
             raise ProductAlreadyExists('the email is already in use')
 
         ProductRepository.update_product(product)
+
+        email_text = cls.get_email_text(user_public_id, old_product, product)
+        EmailHelper.send_email(
+            f'Updated product: {old_product.name}',
+            email_text,
+            to_admins=True,
+            user_public_id=user_public_id
+        )        
 
         return product
 
@@ -75,6 +90,11 @@ class ProductController():
     @classmethod
     def delete_product(cls, public_id: str):
 
+        product = ProductController.get_product_by_public_id(public_id)
+
+        if not product:
+            raise ProductNotFound()
+
         ProductRepository.delete_product(public_id)
 
     @classmethod
@@ -84,3 +104,20 @@ class ProductController():
 
         if (user and user.role == Roles.USER) or not user:
             ProductRepository.increment_queries(public_id)
+
+    @classmethod
+    def get_email_text(cls, user_public_id: str, old_product: ProductData, new_product: ProductData) -> str:
+        user = UserRepository.get_user_by_public_id(user_public_id)
+
+        email_text = f'The user {user.username} has updated the following values of the product {old_product.name}:\r'
+
+        old_product_dict = old_product.dict()
+        new_product_dict = new_product.dict()
+
+        for key in old_product_dict:
+            if key == 'queries':
+                continue
+            if old_product_dict.get(key) != new_product_dict.get(key):
+                email_text += f'{key}: {old_product_dict.get(key)} -> {new_product_dict.get(key)}\r'
+
+        return email_text
